% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_utils.R
\name{ccc_plot}
\alias{ccc_plot}
\title{Plot Concordance Correlation Coefficient}
\usage{
ccc_plot(x, ...)
}
\arguments{
\item{x}{An object returned from \code{compare_methods}.}

\item{...}{Further arguments that may be passed.}
}
\description{
Plot concordance correlation coefficients (CCCs) for parameters from multiple studies
}
