% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_utils.R
\name{compare_methods}
\alias{compare_methods}
\alias{compare_methods.multiverseMPT}
\title{Compare Estimates from Different Methods}
\usage{
compare_methods(x, ...)

\method{compare_methods}{multiverseMPT}(x, ...)
}
\arguments{
\item{x}{A results object from one or multiple multiverse analyses.}

\item{...}{Further arguments that may be passed.}
}
\description{
Returns all the information necessary to easily compare parameter estimates
from different analysis methods.
}
