% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_utils.R
\name{pairs_plot}
\alias{pairs_plot}
\title{Pairs plot}
\usage{
pairs_plot(x, parameter, ...)
}
\arguments{
\item{x}{An object returned from \code{compare_methods}.}

\item{parameter}{Character. Indicating the name of the parameter to be plotted.}

\item{...}{Further arguments that may be passed.}
}
\description{
Create pairs plots of parameter estimates from multiple studies.
}
