% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_parameters.R
\name{rename_parameters}
\alias{rename_parameters}
\alias{rename_parameters.multiverseMPT}
\title{Rename Parameters in an Analysis Object}
\usage{
rename_parameters(x, rename)

\method{rename_parameters}{multiverseMPT}(x, rename = c())
}
\arguments{
\item{x}{An object of class \code{multiverseMPT}.}

\item{rename.}{A named character vector of the form
\code{c(oldname1 = "newname", oldname2 = "newname2", ...)}.}
}
\description{
Renames parameters in an output object. This might be helpful if you used different
parameter names for the same parameter and want to fix this without recomputing.
}
